refactor: split indexedbasemaps from parser, create abstract class to replace genericode
