+++ src/cubie/systemmodels/__init__.py
@@ -4,5 +4,6 @@ This package provides classes and utilities for defining and solving
 ordinary differential equation systems on CUDA devices.
 """
 
-from cubie.systemmodels.systems import ThreeChamberModel, GenericODE, Decays
-__all__ = ["ThreeChamberModel", "GenericODE", "Decays"]
+from cubie.systemmodels.systems import Decays, ThreeChamberModel
+
+__all__ = ["ThreeChamberModel", "Decays"]
