+++ src/cubie/systemmodels/systems/__init__.py
@@ -5,7 +5,6 @@ systems that can be compiled and executed on CUDA devices.
 """
 
 from cubie.systemmodels.systems.decays import Decays
-from cubie.systemmodels.systems.GenericODE import GenericODE
 from cubie.systemmodels.systems.threeCM import ThreeChamberModel
 
-__all__ = ["Decays", "GenericODE", "ThreeChamberModel"]
+__all__ = ["Decays", "ThreeChamberModel"]
